+++ TrashMob/Controllers/UsersController.cs
@@ -3,6 +3,7 @@ namespace TrashMob.Controllers
 {
     using System;
     using System.Linq;
+    using System.Security.Claims;
     using System.Threading.Tasks;
     using Microsoft.AspNetCore.Authorization;
     using Microsoft.AspNetCore.Mvc;
@@ -46,8 +47,13 @@ namespace TrashMob.Controllers
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutUser(Guid id, User user)
+        public async Task<IActionResult> PutUser(User user)
         {
+            if (!ValidateUser(user.NameIdentifier))
+            {
+                return Forbid();
+            }
+
             try
             {
                 var updatedUser = await userRepository.UpdateUser(user).ConfigureAwait(false);
@@ -55,7 +61,7 @@ namespace TrashMob.Controllers
             }
             catch (DbUpdateConcurrencyException)
             {
-                if (!await UserExists(id).ConfigureAwait(false))
+                if (!await UserExists(user.Id).ConfigureAwait(false))
                 {
                     return NotFound();
                 }
@@ -74,8 +80,13 @@ namespace TrashMob.Controllers
         public async Task<IActionResult> PostUser(User user)
         {
             User originalUser;
-            if ((originalUser = await UserExists(user.TenantId, user.UniqueId).ConfigureAwait(false)) != null)
+            if ((originalUser = await UserExists(user.NameIdentifier).ConfigureAwait(false)) != null)
             {
+                if (!ValidateUser(originalUser.NameIdentifier))
+                {
+                    return Forbid();
+                }
+
                 originalUser.City = user.City;
                 originalUser.Country = user.Country;
                 originalUser.Email = user.Email;
@@ -85,7 +96,7 @@ namespace TrashMob.Controllers
                 originalUser.SurName = user.SurName;
                 originalUser.UserName = user.UserName;
                 await userRepository.UpdateUser(originalUser);
-                var returnedUser = await GetUser(user.TenantId, user.UniqueId).ConfigureAwait(false);
+                var returnedUser = await userRepository.GetUserByNameIdentifier(user.NameIdentifier).ConfigureAwait(false);
                 return Ok(returnedUser);
             }
 
@@ -99,6 +110,13 @@ namespace TrashMob.Controllers
         [RequiredScope(Constants.TrashMobWriteScope)]
         public async Task<IActionResult> DeleteUser(Guid id)
         {
+            var user = await userRepository.GetUserByInternalId(id).ConfigureAwait(false);
+
+            if (!ValidateUser(user.NameIdentifier))
+            {
+                return Forbid();
+            }
+
             await userRepository.DeleteUserByInternalId(id).ConfigureAwait(false);
             return NoContent();
         }
@@ -108,14 +126,15 @@ namespace TrashMob.Controllers
             return (await userRepository.GetAllUsers().ConfigureAwait(false)).Any(e => e.Id == id);
         }
 
-        private async Task<User> UserExists(string tenantId, string uniqueId)
+        private async Task<User> UserExists(string nameIdentifier)
         {
-            return (await userRepository.GetAllUsers().ConfigureAwait(false)).FirstOrDefault(e => e.TenantId == tenantId && e.UniqueId == uniqueId);
+            return (await userRepository.GetAllUsers().ConfigureAwait(false)).FirstOrDefault(u => u.NameIdentifier == nameIdentifier);
         }
 
-        private Task<User> GetUser(string tenantId, string uniqueId)
+        private bool ValidateUser(string userId)
         {
-            return userRepository.GetUserByExternalId(tenantId, uniqueId);
+            var nameIdentifier = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            return userId == nameIdentifier;
         }
     }
 }
